Try to get nice markdown formatting; not reliable
The vim-pandoc-syntax *sometimes* loads, and correctly handles YAML
front-matter. It also does nice stuff like folding and conceal.

But it only works sometimes (on my YAML-started frontmatter files). Even
with the same config.

I think I'm going to drop this; it just doesn't work.
[This](https://www.maero.dk/markdown-frontmatter-syntax-highlighting/)
seems to be the better reference.